add edit group event

+++ doctor-event-api/src/main/java/io/terminus/doctor/event/dto/event/group/edit/DoctorChangeGroupEdit.java
@@ -29,7 +29,7 @@ public class DoctorChangeGroupEdit extends BaseGroupEdit implements Serializable
     /**
      * 品种id
      */
-    private Integer breedId;
+    private Long breedId;
 
     private String breedName;
